Firewall: tear down ethereum, fed's ssh
I hit a tipping point with crypto; I'm done. Ethereum validation was a
PITA. It was janky, and needed a lot of polishing (e.g. just getting
"fake" eth to get on the test net was an adventure in stale URLs).

Also, I shut down the port-forward to fed.nono.io because people were
trying to ssh in, and I didn't like it.